add: add visit doctor service impl

+++ src/main/java/org/harvey/respiratory/server/pojo/entity/DiseaseDiagnosisIntermediation.java
@@ -26,5 +26,5 @@ public class DiseaseDiagnosisIntermediation {
     /**
      * 疾病表id
      */
-    private Long diseaseId;
+    private Integer diseaseId;
 }
